Rework UUID handling
Require the UUID to be specified in the playbook and set it on file
system creation instead of expecting users to update their playbooks
with the correct UUIDs after the first run.

Until we can expect the target system to have a version of mkfs.gfs2
that supports setting the UUID, we will have to use tunegfs2 to set it.